Refactor dock-alloc-model into modular submodules
- Split model.rs into dec.rs, err.rs, id.rs, problem.rs, req.rs, sol.rs
- Update imports to use prelude and new module structure - Remove unused
noop operator from meta/oplib - Update usages across crates to use
dock_alloc_model::prelude - Add tests for new modules and conversions
