+++ crates/dock-alloc-model/Cargo.toml
@@ -29,5 +29,5 @@ dock_alloc_core = { path = "../dock-alloc-core" }
 num-traits = "0.2.19"
 rand = "0.9.2"
 rand_distr = "0.5.1"
+statrs = "0.18.0"
 tracing = { version = "0.1.41", features = ["max_level_debug", "release_max_level_warn"] }
-tracing-subscriber = "0.3.20"
